added equation of circle questions
This commit adds questions of the form “Find the equation of the circle with centre ... and radius ....” and “Find the centre and radius of the circle with equation ...”. It includes the unit tests.

It also adds a new helper file: geometry.js. Currently this includes functions that take the centre and radius of a circle, and return the equation in the form (x-a)^2+(y-b)^2=r^2 or x^2-2ax+y^2-2bx=C.