finished mathmo question on circle/circle intersections
This commit covers the case of two intersections, along with pretty printing of answers. It also adds the appropriate unit tests, and corrects those broken in the last commit (by the updated geometry helpers).

Note that mathmo may fail if you’ve used the circle/line intersection question (introduced in commit b48fc41). This had the topicID C44, which has been renamed to C34. Solution is to clear the local storage.